+++ server/config/database.js
@@ -2,5 +2,5 @@ module.exports = Object.freeze({
     HOST: "localhost",
     USER: "root",
     PASSWORD: "",
-    DATABASE: "recipe",
+    DATABASE: "recipes",
 });