+++ server/entities/recipe.js
@@ -1,6 +1,6 @@
-module.exports = class Recipe{
-    
-    constructor(id, name, type, ingredients, book, plate, calories, photo, grade, creationDateTime, lastModificationDateTime){
+module.exports = class Recipe {
+
+    constructor(id, name, type, ingredients, book, plate, calories, photoUrl, photoUrlBackup, grade, lastModificationDateTime, creationDateTime) {
         this.id = id
         this.name = name
         this.type = type
@@ -8,13 +8,14 @@ module.exports = class Recipe{
         this.book = book
         this.plate = plate
         this.calories = calories
-        this.photo = photo
+        this.photoUrl = photoUrl
+        this.photoUrlBackup = photoUrlBackup
         this.grade = grade
-        this.creationDateTime = creationDateTime
         this.lastModificationDateTime = lastModificationDateTime
+        this.creationDateTime = creationDateTime
     }
 
-    print(){
+    print() {
         console.log("id: " + this.id)
         console.log("name: " + this.name)
         console.log("type: " + this.type)
@@ -22,7 +23,8 @@ module.exports = class Recipe{
         console.log("book: " + this.book)
         console.log("plate: " + this.plate)
         console.log("calories: " + this.calories)
-        console.log("photo: " + this.photo)
+        console.log("photoUrl: " + this.photoUrl)
+        console.log("photoUrlBackup: " + this.photoUrlBackup)
         console.log("grade: " + this.grade)
         console.log("createdBy: " + this.createdBy)
         console.log("creationDateTime: " + this.creationDateTime)