+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file is a contains a list of changes for each release.
 
+## v0.2.0 (20-12-2021)
+
+* Add recipes to the database
+* Create database
+* Create `toolbox`
+* Create `Database.sql`
+* Create method to execute queries with parameters
+* Create `index.js`
+* Update `app.js`
+
 ## v0.1.0 (19-12-2021)
 
 * Create `CHANGELOG`